backend/requirements.txt
fastapi
uvicorn
supabase
requests
pydantic
python-dotenv
